refactor(plugins): share manifest capability families
